UI v0.4
New Login and Sign Up page
