+++ tlc-website copy/src/pages/LoginSignUp/LoginSignUp.js
@@ -1,11 +1,13 @@
-import { signInWithPopup, GoogleAuthProvider, signOut, onAuthStateChanged, signInWithEmailAndPassword, createUserWithEmailAndPassword, sendEmailVerification, updateProfile } from "firebase/auth";
-import { auth, storage } from '../../firebase/firebase';
+import React from "react";
+import './LoginSignUp.css';
+import Login from '../../components/UserLogin/Login.js';
+import '../../App.css';
 
 function LoginSignUp(){
 
     return (
         <div className="App-background">
-
+            <Login />
         </div>
     );
 }
